refactor following dan's comments
